Apply SVT-AV1 v1.6.0 patches for ffmpeg n5.1 (#6)
* avcodec/libsvtav1: remove compressed_ten_bit_format and simplify alloc_buffer

compressed_ten_bit_format has been deprecated upstream and has no effect
and can be removed. Plus, technically it was never used in the first place
since it would require the app (ffmpeg) to set it and do additional
processing of the input frames.

Also simplify alloc_buffer by removing calculations relating to the
non-existant processing.

Signed-off-by: Christopher Degawa <christopher.degawa@intel.com>

* avcodec/libsvtav1: replace vbv_bufsize with maximum_buffer_size_ms

svt-av1 v1.2.0 has deprecated vbv_bufsize in favor of using
- maximum_buffer_size_ms (--buf-sz)
- starting_buffer_level_ms (--buf-initial-sz)
- optimal_buffer_level_ms (--buf-optimal-sz)

and vbv_bufsize has not been in use since svt-av1 v0.8.6

Signed-off-by: Christopher Degawa <christopher.degawa@intel.com>

* avcodec/libsvtav1: only set max_buf_sz if both bitrate and rc_buf_sz is set

maximum_buffer_size_ms should only be set if both are specified or if
the user sets it through -svtav1-params buf-sz=val

Signed-off-by: Christopher Degawa <ccom@randomderp.com>

* avcodec/libsvtav1: don't force a default value for deprecated options

Stick to the library's default value instead.

Should address AOMediaCodec/SVT-AV1 issue #2011.

Signed-off-by: James Almer <jamrial@gmail.com>

* avcodec/libsvtav1: use larger of bit rate and max rate for buffer size

Generally if maxrate is set, the calculation should be maxrate over
bufsize. This additionally enables CRF + maxrate & bufsize usage.

In order to keep negative values from enabling zero to be treated
as larger and causing a division by zero, check that one of the
variables is larger than zero.

* libavcodec/libsvtav1: Allow -1 for preset option

Currently, the -1 (MR) preset is disallowed as it's taken as the preset
option not set, and the only way to access it was through svtav1-params.

Signed-off-by: Christopher Degawa <ccom@randomderp.com>
Signed-off-by: James Almer <jamrial@gmail.com>

---------

Signed-off-by: Christopher Degawa <christopher.degawa@intel.com>
Signed-off-by: Christopher Degawa <ccom@randomderp.com>
Signed-off-by: James Almer <jamrial@gmail.com>
Co-authored-by: Christopher Degawa <christopher.degawa@intel.com>
Co-authored-by: Christopher Degawa <ccom@randomderp.com>
Co-authored-by: James Almer <jamrial@gmail.com>
Co-authored-by: Jan Ekström <jeebjp@gmail.com>